Criacao da classe conta para realizar a criacao de uma tabela referente a conta e uma tabela relacionada referente ao cliente

+++ src/main/java/com/desafio/apitransferencia/repository/ClienteRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ClienteRepository extends JpaRepository<Cliente, Long> {
-    Optional <Cliente> findClienteByNumeroConta(int numeroConta);
+    Optional <Cliente> findClienteByNumeroConta(long numeroConta);
 
     List<Cliente> findAll();
 }
